Add guide and few changes
