+++ algo_hs_test.go
@@ -40,19 +40,6 @@ func TestHS(t *testing.T) {
 	f(HS256, hsKey256, hsKeyAnother256, ErrInvalidSignature)
 }
 
-func TestNewKey(t *testing.T) {
-	key, err := GenerateRandom512Bit()
-	if err != nil {
-		t.Fatalf("Error returned directly from GenerateRandom512Bit: %e", err)
-	}
-
-	// 8 bits to 1 byte
-	const byteCount = int(512.0 / 8)
-	if l := len(key); l != byteCount {
-		t.Fatalf("length of key is %d, want %d", l, byteCount)
-	}
-}
-
 var (
 	hsKey256 = []byte("hmac-secret-key-256")
 	hsKey384 = []byte("hmac-secret-key-384")
